reddit_stream/pyproject.toml
[project]
name = "reddit-stream"
version = "0.1.0"
description = "Real-time Reddit streamer for multiple subreddits"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.33.1",
    "pyyaml>=6.0",
    "kafka-python>=2.0.2",
    "pydantic>=2.0.0",
]